+++ mot/cl_routines/mapping/objective_list_calculator.py
@@ -119,16 +119,16 @@ class _ObjectiveListCalculatorWorker(Worker):
             __kernel void get_objectives(
                 ''' + ",\n".join(kernel_param_names) + '''
                 ){
-                    int gid = get_global_id(0);
+                    ulong gid = get_global_id(0);
                     ''' + self._model.get_kernel_data_struct_initialization(self._cl_environment.device, 'data') + '''
 
                     mot_float_type x[''' + str(nmr_params) + '''];
-                    for(int i = 0; i < ''' + str(nmr_params) + '''; i++){
+                    for(uint i = 0; i < ''' + str(nmr_params) + '''; i++){
                         x[i] = params[gid * ''' + str(nmr_params) + ''' + i];
                     }
 
                     global mot_float_type* result = objectives + gid * NMR_INST_PER_PROBLEM;
-                    for(int i = 0; i < NMR_INST_PER_PROBLEM; i++){
+                    for(uint i = 0; i < NMR_INST_PER_PROBLEM; i++){
                         result[i] = getObjectiveInstanceValue((void*)&data, x, i);
                     }
             }
